# app/static/swagger.json
{
...
  "info": {
    "description": "A RESTful API for a sports team and player matching platform.",
    "title": "Sport Matching API",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/teams": {
      "get": {
        "responses": {
          "200": {
            "description": "A list of teams."
          }
        },
        "summary": "Get all teams",
        "tags": [
          "Teams"
        ]
      },
      "post": {
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "city": {
                    "type": "string"
                  },
                  "logoUrl": {
                    "type": "string"
                  },
                  "name": {
                    "type": "string"
                  },
                  "sport_id": {
                    "type": "string"
                  }
                },
                "required": [
                  "name",
                  "sport_id",
                  "city"
                ],
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Team created successfully."
          },
          "400": {
            "description": "Missing required fields."
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Create a new team",
        "tags": [
          "Teams"
        ]
      }
    },
    "/api/v1/teams/{team_id}": {
      "get": {
        "parameters": [
          {
            "in": "path",
            "name": "team_id",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "A single team."
          },
          "404": {
            "description": "Team not found."
          }
        },
        "summary": "Get a team by ID",
        "tags": [
          "Teams"
        ]
      }
    },
    "/api/v1/teams/{team_id}/applications": {
      "get": {
        "parameters": [
          {
            "in": "path",
            "name": "team_id",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "A list of user applications."
          },
          "403": {
            "description": "Forbidden."
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Get team applications",
        "tags": [
          "Teams"
        ]
      }
    },
    "/api/v1/teams/{team_id}/applications/{user_id}/respond": {
      "post": {
        "parameters": [
          {
            "in": "path",
            "name": "team_id",
            "required": true,
            "type": "integer"
          },
          {
            "in": "path",
            "name": "user_id",
            "required": true,
            "type": "integer"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "action": {
                    "enum": [
                      "accept",
                      "decline"
                    ],
                    "type": "string"
                  }
                },
                "required": [
                  "action"
                ],
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Response recorded."
          },
          "400": {
            "description": "Invalid action."
          },
          "403": {
            "description": "Forbidden."
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Respond to a team application",
        "tags": [
          "Teams"
        ]
      }
    },
    "/api/v1/teams/{team_id}/apply": {
      "post": {
        "parameters": [
          {
            "in": "path",
            "name": "team_id",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "201": {
            "description": "Application sent successfully."
          },
          "409": {
            "description": "Application already sent."
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Apply to a team",
        "tags": [
          "Teams"
        ]
      }
    },
    "/api/v1/teams/{team_id}/follow": {
      "post": {
        "parameters": [
          {
            "in": "path",
            "name": "team_id",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "Follow status toggled."
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Toggle follow a team",
        "tags": [
          "Teams"
        ]
      }
    },
    "/api/v1/teams/{team_id}/logo": {
      "post": {
        "parameters": [
          {
            "in": "path",
            "name": "team_id",
            "required": true,
            "type": "integer"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "fileUrl": {
                    "type": "string"
                  }
                },
                "required": [
                  "fileUrl"
                ],
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Logo updated successfully."
          },
          "400": {
            "description": "Missing fileUrl."
          },
          "403": {
            "description": "Only the captain can update."
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Update team logo",
        "tags": [
          "Teams"
        ]
      }
    },
    "/api/v1/teams/{team_id}/members/{user_id}": {
      "delete": {
        "parameters": [
          {
            "in": "path",
            "name": "team_id",
            "required": true,
            "type": "integer"
          },
          {
            "in": "path",
            "name": "user_id",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "Player removed successfully."
          },
          "400": {
            "description": "Captain cannot remove themselves."
          },
          "403": {
            "description": "Only the team captain can remove members."
          },
          "404": {
            "description": "Player is not a member of this team."
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Remove a team member",
        "tags": [
          "Teams"
        ]
      }
    },
...
  },
  "schemes": [
    "http",
    "https"
  ],
  "securityDefinitions": {
    "Bearer": {
      "description": "JWT Authorization header using the Bearer scheme. Example: \"Authorization: Bearer {token}\"",
      "in": "header",
      "name": "Authorization",
      "type": "apiKey"
    }
  },
  "swagger": "2.0"
}